fix(normalization): accept float percents and rescale fractions to 0-100
LLM returned target_cogs_pct=0.3 (fraction) but schema required int,
crashing the job. Changed target_cogs_pct + sales_mix_*_pct fields
from int to float, and added coercion in normalize_intake that rescales
any percent value in (0, 1] to its 0–100 equivalent before validation.
Covers target_cogs_pct, sales_mix_*, funding_*, revenue_growth, ramp_start.

+++ schemas/concept_schema.py
@@ -65,11 +65,11 @@ class ConceptObject(BaseModel):
 
     staff_model: Optional[Literal["lean", "standard", "full", "custom"]] = None
 
-    sales_mix_dinein_pct: Optional[int] = None
-    sales_mix_takeaway_pct: Optional[int] = None
-    sales_mix_delivery_pct: Optional[int] = None
+    sales_mix_dinein_pct: Optional[float] = None
+    sales_mix_takeaway_pct: Optional[float] = None
+    sales_mix_delivery_pct: Optional[float] = None
 
-    target_cogs_pct: Optional[int] = None
+    target_cogs_pct: Optional[float] = None
 
     kitchen_type: Optional[Literal["full_line", "prep_finish", "assembly_only", "central_kitchen"]] = None
     operating_days_per_week: Optional[int] = None
